Working payments calculator

+++ app/controllers/application_controller.rb
@@ -37,7 +37,7 @@ class ApplicationController < ActionController::Base
   def calculate_monthly_payment
     @apr = params.fetch("user_apr").to_f
     @rate = (params.fetch("user_apr").to_f) / 100 / 12
-    @no_of_years = params.fetch("user_years").to_f
+    @no_of_years = params.fetch("user_years").to_i
     @no_of_periods = params.fetch("user_years").to_f * 12
     @principal = params.fetch("user_pv").to_f
